+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/background/database/BackgroundJob.kt
@@ -2,15 +2,19 @@ package com.outsystems.plugins.healthfitnesslib.background.database
 
 import androidx.room.*
 import androidx.room.ColumnInfo
+import java.util.*
 
 @Entity(
     tableName = BackgroundJob.TABLE_NAME,
-    primaryKeys = ["variable", "comparison", "value"],
     foreignKeys = [ForeignKey(
         entity = Notification::class,
         parentColumns = arrayOf("id"),
         childColumns = arrayOf("notification_id"),
-        onDelete = ForeignKey.CASCADE)]
+        onDelete = ForeignKey.CASCADE)],
+    indices = [Index(
+        "variable", "comparison", "value",
+        name="unique_bg_job",
+        unique=true)]
 )
 open class BackgroundJob {
 
@@ -22,22 +26,11 @@ open class BackgroundJob {
         EQUALS("EQUAL"),
     }
 
+    @PrimaryKey
+    @ColumnInfo(name = "id") var id: String = UUID.randomUUID().toString()
     @ColumnInfo(name = "variable") var variable: String = ""
-        set(value){
-            field = value
-            computeId()
-        }
     @ColumnInfo(name = "comparison") var comparison: String = ""
-        set(value){
-            field = value
-            computeId()
-        }
     @ColumnInfo(name = "value") var value: Float = 0.0f
-        set(value){
-            field = value
-            computeId()
-        }
-
     @ColumnInfo(name = "time_unit") var timeUnit: String? = null
     @ColumnInfo(name = "time_unit_grouping") var timeUnitGrouping: Int? = null
     @ColumnInfo(name = "notification_id") var notificationId: Long? = null
@@ -46,14 +39,7 @@ open class BackgroundJob {
     @ColumnInfo(name = "next_notification_timestamp") var nextNotificationTimestamp: Long = 0
     @ColumnInfo(name = "isActive") var isActive: Boolean = true
 
-    @Ignore
-    var id: String = ""
-
-    private fun computeId() {
-        id = "${this.variable}-${this.comparison}-${this.value}".toLowerCase()
-    }
-
     companion object {
-        const val TABLE_NAME: String= "backgroundJob"
+        const val TABLE_NAME: String= "backgroundJob_2"
     }
 }