+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/background/database/DatabaseManager.kt
@@ -52,8 +52,8 @@ class DatabaseManager(context : Context) : DatabaseManagerInterface {
         return notificationDao?.getAll()
     }
 
-    override fun fetchBackgroundJob(variable: String, comparison: String, value: Float) : BackgroundJob? {
-        return backgroundJobDao?.findByPrimaryKey(variable, comparison, value)
+    override fun fetchBackgroundJob(id: String) : BackgroundJob? {
+        return backgroundJobDao?.findById(id)
     }
 
     override fun fetchBackgroundJobCountForVariable(variable: String) : Int {
@@ -78,6 +78,10 @@ class DatabaseManager(context : Context) : DatabaseManagerInterface {
 
     override fun updateBackgroundJob(backgroundJob: BackgroundJob) {
         backgroundJobDao?.update(backgroundJob)
+    }
+
+    override fun updateNotification(notification: Notification) {
+        notificationDao?.update(notification)
     }
 
     override fun runInTransaction(closude : () -> Unit){